requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.6.0
pypdf>=4.0.0
langchain-google-genai
langchain-community
langchain-text-splitters
google-generativeai>=0.8.0


